move all files to source and fix few warnings and includes
