+++ .clangd
@@ -1,0 +1,9 @@
+CompileFlags:
+  Add: [-std=c++20]
+
+Diagnostics:
+  UnusedIncludes: Strict
+  # MissingIncludes: Strict
+
+Hover:
+  ShowAKA: true
