+++ source/common/CMakeLists.txt
@@ -2,7 +2,7 @@ add_library(common INTERFACE)
 
 target_include_directories(common
     INTERFACE
-        .
+        include
 )
 
 target_link_libraries(common
