+++ source/engine/CMakeLists.txt
@@ -1,10 +1,9 @@
 add_library(engine STATIC)
 
-add_subdirectory(atoms_list)
-
 target_sources(engine
     PRIVATE
         source/engine.cpp
+        source/eng_atoms_list.cpp
 )
 
 target_include_directories(engine
@@ -14,8 +13,8 @@ target_include_directories(engine
 
 target_link_libraries(engine
     PRIVATE
+        common
         flags
-        atoms_list
         libs
         glm
 )
