+++ source/graphics/include/graphics/graphics_defs.h
